requirements.txt
django>=4.2.0
django-ninja>=1.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
celery>=5.3.0
django-cors-headers>=4.0.0
python-decouple>=3.8
aiohttp>=3.8.0
requests>=2.28.0